fix: Motion sensor activation overrides state
This fixes the following scenario: say I have a motion sensor that has
activated and then I adjust the brightness and then leave the room
temporarily and re-enter. Previously, the manually-state brightness
would be overridden, since we would reset the state when the sensor is
activated again.

After this change, we only attempt to set the scene/turn the light on if
it is currently off. This avoids a situation where we are overwriting
any state that the user has set in the mean time.