fix: Bug where lights turn off unexpectedly
This fixes the following issue: user is in the room and the lights dim
because motion has not been triggered for a while. The user presses the
"on" button to say "no, I don't want the lights to turn off". Before
this change, the lights would become brighter again but the turn off
timer wouldn't be cancelled so the lights would turn off anyway.

After this change, the turn off timer will be cleared when the lights
are changed by a human so they won't unexpectedly turn off when the user
has just turned them back on/brightness up.