fix: Light not turning off after dimming
Turns out hal does get state updates from service/actions that it
triggers. Ideally we need a way for hal to know that a state update was
triggered by an action on our side, e.g. by passing an ID or context
somehow. I did some brief research and it doesn't look like it's
possible. Therefore until we have a better way, when we publish a
message we just increment a counter for that entity and assume that the
next state update is due to our action. If the counter is > 0, we don't
trigger automations. This prevents loops, i.e. automations triggering an
action which then re-trigger the automation.